+++ Bencodex.Tests/Types/ListTest.cs
@@ -5,6 +5,8 @@ using Bencodex.Types;
 using Xunit;
 using static Bencodex.Misc.ImmutableByteArrayExtensions;
 using static Bencodex.Tests.TestUtils;
+using IEquatableValues =
+    System.IEquatable<System.Collections.Immutable.IImmutableList<Bencodex.Types.IValue>>;
 
 namespace Bencodex.Tests.Types
 {
@@ -28,11 +30,6 @@ namespace Bencodex.Tests.Types
         [Fact]
         public void Constructors()
         {
-            Assert.Equal(
-                _one.Add(_zero).Add(_one).Add(_two),
-                new List(ImmutableArray.Create<IValue>(Null.Value, _zero, _one, _two))
-            );
-
             Assert.Equal(_zero, new List(Enumerable.Empty<IValue>())
             );
             Assert.Equal(
@@ -158,6 +155,57 @@ namespace Bencodex.Tests.Types
             Assert.Equal(Enumerable.Count(_zero), _zero.Count);
             Assert.Equal(Enumerable.Count(_one), _one.Count);
             Assert.Equal(Enumerable.Count(_two), _two.Count);
+        }
+
+        [Fact]
+        public void Enumerate()
+        {
+            Assert.Empty(_zero.ToArray());
+            Assert.Equal(new IValue[] { Null.Value }, _one.ToArray());
+            Assert.Equal(
+                new Text[] { "hello", "world" }.Cast<IValue>(),
+                _two.ToArray()
+            );
+            Assert.Equal(
+                new IValue[] { Null.Value, _zero, _one, _two },
+                _nest.ToArray()
+            );
+        }
+
+        [Fact]
+        public void Equality()
+        {
+            Assert.True(_zero.Equals(new List()));
+            Assert.True(((IEquatableValues)_zero).Equals(ImmutableArray<IValue>.Empty));
+            Assert.True(_one.Equals(new List(Null.Value)));
+            Assert.True(
+                ((IEquatableValues)_one).Equals(ImmutableArray<IValue>.Empty.Add(Null.Value))
+            );
+            Assert.True(_two.Equals(new List((Text)"hello", (Text)"world")));
+            Assert.True(
+                ((IEquatableValues)_two).Equals(
+                    ImmutableArray.Create<IValue>((Text)"hello", (Text)"world")
+                )
+            );
+            Assert.True(_nest.Equals(new List(Null.Value, _zero, _one, _two)));
+            Assert.True(
+                ((IEquatableValues)_nest).Equals(
+                    ImmutableArray.Create<IValue>(Null.Value, _zero, _one, _two)
+                )
+            );
+
+            Assert.False(_zero.Equals(_one));
+            Assert.False(_zero.Equals(_two));
+            Assert.False(_zero.Equals(_nest));
+            Assert.False(_one.Equals(_zero));
+            Assert.False(_one.Equals(_two));
+            Assert.False(_one.Equals(_nest));
+            Assert.False(_two.Equals(_zero));
+            Assert.False(_two.Equals(_one));
+            Assert.False(_two.Equals(_nest));
+            Assert.False(_nest.Equals(_one));
+            Assert.False(_nest.Equals(_two));
+            Assert.False(_nest.Equals(_two));
         }
 
         [Fact]