+++ Bencodex/Types/List.cs
@@ -7,16 +7,19 @@ using System.IO;
 using System.Linq;
 using System.Numerics;
 using System.Security.Cryptography;
+using Bencodex.Misc;
 
 namespace Bencodex.Types
 {
     /// <summary>
     /// Represents Bencodex lists.
     /// </summary>
+    [Pure]
     public class List :
         IValue,
         IImmutableList<IValue>,
-        IEquatable<IImmutableList<IValue>>
+        IEquatable<IImmutableList<IValue>>,
+        IEquatable<List>
     {
         /// <summary>
         /// The empty list.
@@ -31,7 +34,8 @@ namespace Bencodex.Types
 
         private static readonly byte[] _listPrefix = new byte[1] { 0x6c };  // 'l'
 
-        private ImmutableArray<IValue> _value;
+        private ImmutableArray<IndirectValue> _values;
+        private IndirectValue.Loader? _loader;
         private long? _encodingLength;
         private ImmutableArray<byte>? _hash;
 
@@ -40,7 +44,7 @@ namespace Bencodex.Types
         /// </summary>
         /// <param name="elements">The element values to include.</param>
         public List(params IValue[] elements)
-            : this(ImmutableArray.Create(elements))
+            : this((IEnumerable<IValue>)elements)
         {
         }
 
@@ -49,17 +53,34 @@ namespace Bencodex.Types
         /// </summary>
         /// <param name="elements">The element values to include.</param>
         public List(IEnumerable<IValue> elements)
-            : this(elements.ToImmutableArray())
+            : this(elements.Select(v => new IndirectValue(v)).ToImmutableArray(), null)
         {
         }
 
         /// <summary>
-        /// Creates a <see cref="List"/> instance with <paramref name="elements"/>.
+        /// Creates a <see cref="List"/> instance with <paramref name="indirectValues"/> and
+        /// a <paramref name="loader"/> used for loading unloaded values.
         /// </summary>
-        /// <param name="elements">The element values to include.</param>
-        public List(in ImmutableArray<IValue> elements)
+        /// <param name="indirectValues">The loaded and unloaded values to include.</param>
+        /// <param name="loader">The <see cref="IndirectValue.Loader"/> delegate invoked when
+        /// unloaded values are needed.</param>
+        public List(IEnumerable<IndirectValue> indirectValues, IndirectValue.Loader loader)
+            : this(
+                indirectValues is ImmutableArray<IndirectValue> ia
+                    ? ia
+                    : indirectValues.ToImmutableArray(),
+                loader
+            )
         {
-            _value = elements;
+        }
+
+        private List(
+            in ImmutableArray<IndirectValue> indirectValues,
+            IndirectValue.Loader? loader = null
+        )
+        {
+            _values = indirectValues;
+            _loader = loader;
             _encodingLength = null;
             _hash = null;
         }
@@ -74,7 +95,7 @@ namespace Bencodex.Types
         {
             get
             {
-                if (_value.IsDefaultOrEmpty)
+                if (_values.IsDefaultOrEmpty)
                 {
                     return EmptyFingerprint;
                 }
@@ -84,7 +105,7 @@ namespace Bencodex.Types
                     long encLength = 2;
                     SHA1 sha1 = SHA1.Create();
                     sha1.Initialize();
-                    foreach (IValue value in _value)
+                    foreach (IndirectValue value in _values)
                     {
                         Fingerprint fp = value.Fingerprint;
                         byte[] fpb = fp.Serialize();
@@ -106,11 +127,12 @@ namespace Bencodex.Types
         }
 
         /// <inheritdoc cref="IValue.EncodingLength"/>
-        [Pure]
         public long EncodingLength =>
             _encodingLength is { } l ? l : (
                 _encodingLength = _listPrefix.LongLength
-                    + _value.Sum(e => e.EncodingLength)
+                + _values.Sum(e => e.LoadedValue is { } v
+                        ? v.EncodingLength
+                        : e.Fingerprint.EncodingLength)
                     + CommonVariables.Suffix.LongLength
             ).Value;
 
@@ -118,26 +140,57 @@ namespace Bencodex.Types
         [Obsolete("Deprecated in favour of " + nameof(Inspect) + "() method.")]
         public string Inspection => Inspect(true);
 
-        public int Count => _value.Length;
+        /// <inheritdoc cref="IReadOnlyCollection{T}.Count"/>
+        public int Count => _values.Length;
 
-        public IValue this[int index] => _value[index];
+        /// <inheritdoc cref="IReadOnlyList{T}.this[int]"/>
+        public IValue this[int index] => _values[index].GetValue(_loader);
 
-        bool IEquatable<IImmutableList<IValue>>.Equals(
-            IImmutableList<IValue> other
-        )
+        bool IEquatable<IImmutableList<IValue>>.Equals(IImmutableList<IValue> other)
         {
-            return _value.SequenceEqual(other);
+            if (Count != other.Count)
+            {
+                return false;
+            }
+            else if (_hash is { } && other is List otherList && otherList._hash is { })
+            {
+                return Fingerprint.Equals(otherList.Fingerprint);
+            }
+
+            for (int i = 0; i < _values.Length; i++)
+            {
+                IndirectValue iv = _values[i];
+                IValue ov = other[i];
+                if (iv.LoadedValue is { } v)
+                {
+                    if (!ov.Equals(v))
+                    {
+                        return false;
+                    }
+
+                    continue;
+                }
+
+                if (!iv.Fingerprint.Equals(ov.Fingerprint))
+                {
+                    return false;
+                }
+            }
+
+            return true;
         }
 
+        public bool Equals(List other) =>
+            ((IEquatable<IImmutableList<IValue>>)this).Equals(other);
+
         bool IEquatable<IValue>.Equals(IValue other) =>
-            other is List o &&
-            ((IEquatable<IImmutableList<IValue>>)this).Equals(o);
+            other is List o && Equals(o);
 
         IEnumerator<IValue> IEnumerable<IValue>.GetEnumerator()
         {
-            foreach (IValue element in _value)
+            foreach (IndirectValue element in _values)
             {
-                yield return element;
+                yield return element.GetValue(_loader);
             }
         }
 
@@ -153,147 +206,157 @@ namespace Bencodex.Types
         }
 
         /// <inheritdoc cref="object.GetHashCode()"/>
-        [Pure]
         public override int GetHashCode() =>
-            _value.GetHashCode();
+            _values.GetHashCode();
 
-        IEnumerator IEnumerable.GetEnumerator()
-        {
-            return ((IEnumerable)_value).GetEnumerator();
-        }
+        IEnumerator IEnumerable.GetEnumerator() =>
+            ((IEnumerable<IValue>)this).GetEnumerator();
 
-        IImmutableList<IValue> IImmutableList<IValue>.Add(IValue value)
-        {
-            return new List(_value.Add(value));
-        }
+        IImmutableList<IValue> IImmutableList<IValue>.Add(IValue value) =>
+            Add(value);
 
-        public List Add(IValue value)
-        {
-            return new List(_value.Add(value));
-        }
+        /// <summary>Makes a copy of the list, and adds the specified <paramref name="value"/>
+        /// to the end of the copied list.</summary>
+        /// <param name="value">The value to add to the list.</param>
+        /// <returns>A new list with the value added.</returns>
+        public List Add(IValue value) =>
+            new List(_values.Add(new IndirectValue(value)));
 
-        public List Add(string value)
-        {
-            return new List(_value.Add((Text)value));
-        }
+        /// <summary>Makes a copy of the list, and adds the specified <paramref name="value"/>
+        /// to the end of the copied list.</summary>
+        /// <param name="value">The value to add to the list.  It is automatically turned into
+        /// a Bencodex <see cref="Text"/> instance.</param>
+        /// <returns>A new list with the value added.</returns>
+        public List Add(string value) =>
+            Add((IValue)new Text(value));
 
-        public List Add(bool value)
-        {
-            return new List(_value.Add((Boolean)value));
-        }
+        /// <summary>Makes a copy of the list, and adds the specified <paramref name="value"/>
+        /// to the end of the copied list.</summary>
+        /// <param name="value">The value to add to the list.  It is automatically turned into
+        /// a Bencodex <see cref="Boolean"/> instance.</param>
+        /// <returns>A new list with the value added.</returns>
+        public List Add(bool value) =>
+            Add((IValue)new Boolean(value));
 
-        public List Add(BigInteger value)
-        {
-            return new List(_value.Add((Integer)value));
-        }
+        /// <summary>Makes a copy of the list, and adds the specified <paramref name="value"/>
+        /// to the end of the copied list.</summary>
+        /// <param name="value">The value to add to the list.  It is automatically turned into
+        /// a Bencodex <see cref="Integer"/> instance.</param>
+        /// <returns>A new list with the value added.</returns>
+        public List Add(BigInteger value) =>
+            Add((IValue)new Integer(value));
 
-        public List Add(byte[] value)
-        {
-            return new List(_value.Add((Binary)value));
-        }
+        /// <summary>Makes a copy of the list, and adds the specified <paramref name="value"/>
+        /// to the end of the copied list.</summary>
+        /// <param name="value">The value to add to the list.  It is automatically turned into
+        /// a Bencodex <see cref="Binary"/> instance.</param>
+        /// <returns>A new list with the value added.</returns>
+        public List Add(byte[] value) =>
+            Add((IValue)new Binary(value));
 
         IImmutableList<IValue> IImmutableList<IValue>.AddRange(
             IEnumerable<IValue> items
-        )
-        {
-            return new List(_value.AddRange(items));
-        }
+        ) =>
+            new List(_values.AddRange(items.Select(v => new IndirectValue(v))));
 
-        IImmutableList<IValue> IImmutableList<IValue>.Clear()
-        {
-            return new List(_value.Clear());
-        }
+        IImmutableList<IValue> IImmutableList<IValue>.Clear() =>
+            List.Empty;
 
+        [Obsolete("This operation immediately loads all unloaded values in the range on " +
+                  "the memory.")]
         int IImmutableList<IValue>.IndexOf(
             IValue item,
             int index,
             int count,
             IEqualityComparer<IValue> equalityComparer
-        )
-        {
-            return _value.IndexOf(item, index, count, equalityComparer);
-        }
+        ) =>
+            _values.IndexOf(
+                new IndirectValue(item),
+                index,
+                count,
+                new IndirectValueEqualityComparer(equalityComparer, _loader)
+            );
 
-        IImmutableList<IValue> IImmutableList<IValue>.Insert(
-            int index,
-            IValue element
-        )
-        {
-            return new List(_value.Insert(index, element));
-        }
+        IImmutableList<IValue> IImmutableList<IValue>.Insert(int index, IValue element) =>
+            new List(_values.Insert(index, new IndirectValue(element)));
 
         IImmutableList<IValue> IImmutableList<IValue>.InsertRange(
             int index,
             IEnumerable<IValue> items
         )
         {
-            return new List(_value.InsertRange(index, items));
+            IEnumerable<IndirectValue> vs = items.Select(v => new IndirectValue(v));
+            return new List(_values.InsertRange(index, vs));
         }
 
+        [Obsolete("This operation immediately loads all unloaded values in the range on " +
+                  "the memory.")]
         int IImmutableList<IValue>.LastIndexOf(
             IValue item,
             int index,
             int count,
             IEqualityComparer<IValue> equalityComparer
-        )
-        {
-            return _value.LastIndexOf(item, index, count, equalityComparer);
-        }
+        ) =>
+            _values.LastIndexOf(
+                new IndirectValue(item),
+                index,
+                count,
+                new IndirectValueEqualityComparer(equalityComparer, _loader)
+            );
 
+        [Obsolete("This operation immediately loads all unloaded values on the memory.")]
         IImmutableList<IValue> IImmutableList<IValue>.Remove(
             IValue value,
             IEqualityComparer<IValue> equalityComparer
-        )
-        {
-            return new List(_value.Remove(value, equalityComparer));
-        }
+        ) =>
+            new List(
+                _values.Remove(
+                    new IndirectValue(value),
+                    new IndirectValueEqualityComparer(equalityComparer, _loader)
+                ),
+                _loader
+            );
 
+        [Obsolete("This operation immediately loads all unloaded values on the memory.")]
         IImmutableList<IValue> IImmutableList<IValue>.RemoveAll(
             Predicate<IValue> match
-        )
-        {
-            return new List(_value.RemoveAll(match));
-        }
+        ) =>
+            new List(_values.RemoveAll(iv => match(iv.GetValue(_loader))), _loader);
 
-        IImmutableList<IValue> IImmutableList<IValue>.RemoveAt(int index)
-        {
-            return new List(_value.RemoveAt(index));
-        }
+        IImmutableList<IValue> IImmutableList<IValue>.RemoveAt(int index) =>
+            new List(_values.RemoveAt(index));
 
+        [Obsolete("This operation immediately loads all unloaded values on the memory.")]
         IImmutableList<IValue> IImmutableList<IValue>.RemoveRange(
             IEnumerable<IValue> items,
             IEqualityComparer<IValue> equalityComparer
-        )
-        {
-            return new List(_value.RemoveRange(items, equalityComparer));
-        }
+        ) =>
+            new List(
+                _values.RemoveRange(
+                    items.Select(v => new IndirectValue(v)),
+                    new IndirectValueEqualityComparer(equalityComparer, _loader)
+                )
+            );
 
-        IImmutableList<IValue> IImmutableList<IValue>.RemoveRange(
-            int index,
-            int count
-        )
-        {
-            return new List(_value.RemoveRange(index, count));
-        }
+        IImmutableList<IValue> IImmutableList<IValue>.RemoveRange(int index, int count) =>
+            new List(_values.RemoveRange(index, count));
 
+        [Obsolete("This operation immediately loads all unloaded values on the memory.")]
         IImmutableList<IValue> IImmutableList<IValue>.Replace(
             IValue oldValue,
             IValue newValue,
             IEqualityComparer<IValue> equalityComparer
-        )
-        {
-            return new List(
-                _value.Replace(oldValue, newValue, equalityComparer)
+        ) =>
+            new List(
+                _values.Replace(
+                    new IndirectValue(oldValue),
+                    new IndirectValue(newValue),
+                    new IndirectValueEqualityComparer(equalityComparer, _loader)
+                )
             );
-        }
 
-        IImmutableList<IValue> IImmutableList<IValue>.SetItem(
-            int index,
-            IValue value
-        )
-        {
-            return new List(_value.SetItem(index, value));
-        }
+        IImmutableList<IValue> IImmutableList<IValue>.SetItem(int index, IValue value) =>
+            new List(_values.SetItem(index, new IndirectValue(value)));
 
         [Pure]
         public IEnumerable<byte[]> EncodeIntoChunks()
@@ -314,11 +377,12 @@ namespace Bencodex.Types
             _encodingLength = length;
         }
 
+        /// <inheritdoc cref="IValue.EncodeToStream(Stream)"/>
         public void EncodeToStream(Stream stream)
         {
             long startPos = stream.Position;
             stream.WriteByte(_listPrefix[0]);
-            foreach (IValue element in _value)
+            foreach (IValue element in this)
             {
                 element.EncodeToStream(stream);
             }
@@ -330,23 +394,28 @@ namespace Bencodex.Types
         /// <inheritdoc cref="IValue.Inspect(bool)"/>
         public string Inspect(bool loadAll)
         {
-            switch (_value.Length)
+            string InspectItem(IndirectValue value, bool load) =>
+                loadAll || value.LoadedValue is { }
+                    ? value.GetValue(_loader).Inspect(load)
+                    : value.Fingerprint.ToString();
+
+            switch (_values.Length)
             {
                 case 0:
                     return "[]";
 
                 case 1:
-                    var el = this.First();
-                    if (el is List || el is Dictionary)
+                    IndirectValue first = _values[0];
+                    if (first.Type == ValueType.List || first.Type == ValueType.Dictionary)
                     {
                         goto default;
                     }
 
-                    return $"[{el.Inspect(loadAll)}]";
+                    return $"[{InspectItem(first, loadAll)}]";
 
                 default:
-                    IEnumerable<string> elements = this.Select(v =>
-                        $"  {v.Inspect(loadAll).Replace("\n", "\n  ")},\n"
+                    IEnumerable<string> elements = _values.Select(v =>
+                        $"  {InspectItem(v, loadAll).Replace("\n", "\n  ")},\n"
                     );
                     return $"[\n{string.Join(string.Empty, elements)}]";
             }